improve static array initialization
